+++ solvis/filter/parent_fault_id_filter.py
@@ -13,7 +13,7 @@ Examples:
     ```py
     >>> model = InversionSolution.from_archive(filename).model
     >>> parent_fault_ids = FilterParentFaultIds(model)\
-            .for_parent_fault_names(['Alpine Jacksons to Kaniere', 'BooBoo'])
+            .for_parent_fault_names(['Alpine: Jacksons to Kaniere', 'BooBoo'])
     ```
 
 TODO:
@@ -81,7 +81,7 @@ class FilterParentFaultIds:
         ```py
         >>> solution = InversionSolution.from_archive(filename)
         >>> parent_fault_ids = FilterParentFaultIds(solution)\
-                .for_parent_fault_names(['Alpine Jacksons to Kaniere'])
+                .for_parent_fault_names(['Alpine: Jacksons to Kaniere'])
         ```
     """